+++ src/interface/index.js
@@ -1,4 +1,5 @@
 
+export * from "./changeVoiWindow";
 export * from "./initializeElement";
 export * from "./jumpToSlice";
 export * from "./tools";